Django/requirements.txt
asgiref==3.7.2
Django==4.2.5
paho-mqtt==1.6.1
sqlparse==0.4.4
typing_extensions==4.8.0
